Updated skeleton with some pseudo code and comments

+++ src/Profit.java
@@ -16,6 +16,6 @@ public class Profit
 
 	public float globalMaximum() 
 	{
-		return ((2.7 + 0.3 * this.a) / (2.0 - 0.6 * this.b));
+		return (float) ((2.7 + 0.3 * this.a) / (2.0 - 0.6 * this.b));
 	} // globalMaximum
 } // class Profit
